stockprice/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.9
django-filter==2.4.0
djangorestframework==3.12.1
idna==2.10
Markdown==3.3.1
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
